requirements.txt
google-generativeai
sentence-transformers
langchain-community
langchain-core
langchain-google-genai
langgraph
pymupdf
python-dotenv
typing-extensions 